refactor: UserUtil 클래스와 메서드를 역할에 맞게 변경
기존의 클래스명에 맞지 않는 기능들을 재배치
- findById, findByEmail를 역할에 맞게 service로 이동
- convertUserResponse 메서드의 역할에 맞게 클래스명 변경
  - 변경된 클래스 명: UserResponseMapper
