+++ backend/src/main/java/com/github/dadogk/group/util/GroupUtil.java
@@ -2,7 +2,7 @@ package com.github.dadogk.group.util;
 
 import com.github.dadogk.group.dto.create.GroupResponse;
 import com.github.dadogk.group.entity.Group;
-import com.github.dadogk.user.util.UserUtil;
+import com.github.dadogk.user.mapper.UserResponseMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
 
@@ -10,11 +10,12 @@ import org.springframework.stereotype.Component;
 @Component
 public class GroupUtil {
 
-  private final UserUtil userUtil;
+  private final UserResponseMapper userResponseMapper;
 
   public GroupResponse convertGroup(Group group) {
     return new GroupResponse(group.getId(), group.getGroupName(), group.getGroupIntro(),
-        userUtil.convertUserResponse(group.getHostUser()), group.getState(), group.getType(),
+        userResponseMapper.convertUserResponse(group.getHostUser()), group.getState(),
+        group.getType(),
         group.isPrivacyState(), group.getCreatedAt(), group.getGroupMembers().size());
   }
 }
