+++ backend/src/main/java/com/github/dadogk/socket/WebSocketConfiguration.java
@@ -3,7 +3,8 @@ package com.github.dadogk.socket;
 import com.github.dadogk.config.jwt.TokenProvider;
 import com.github.dadogk.study.StudyConnectionHandler;
 import com.github.dadogk.study.StudyService;
-import com.github.dadogk.user.util.UserUtil;
+import com.github.dadogk.user.UserService;
+import com.github.dadogk.user.mapper.UserResponseMapper;
 import lombok.RequiredArgsConstructor;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.web.socket.config.annotation.EnableWebSocket;
@@ -15,13 +16,15 @@ import org.springframework.web.socket.config.annotation.WebSocketHandlerRegistry
 @EnableWebSocket
 public class WebSocketConfiguration implements WebSocketConfigurer {
 
+  private final UserService userService;
   private final StudyService studyService;
   private final TokenProvider tokenProvider;
-  private final UserUtil userUtil;
+  private final UserResponseMapper userResponseMapper;
 
   @Override
   public void registerWebSocketHandlers(WebSocketHandlerRegistry registry) {
-    registry.addHandler(new StudyConnectionHandler(studyService, tokenProvider, userUtil),
+    registry.addHandler(
+        new StudyConnectionHandler(userService, studyService, tokenProvider, userResponseMapper),
         "/study/connect");
   }
 }
