+++ backend/src/main/java/com/github/dadogk/study/StudyApiController.java
@@ -7,8 +7,8 @@ import com.github.dadogk.study.dto.api.recode.RecodeResponse;
 import com.github.dadogk.study.entity.StudyRecord;
 import com.github.dadogk.study.entity.StudySubject;
 import com.github.dadogk.study.util.StudyUtil;
+import com.github.dadogk.user.UserService;
 import com.github.dadogk.user.entity.User;
-import com.github.dadogk.user.util.UserUtil;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
@@ -27,9 +27,9 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("/api/study")
 public class StudyApiController {
 
+  private final UserService userService;
   private final StudyService studyService;
   private final StudyUtil studyUtil;
-  private final UserUtil userUtil;
 
   @GetMapping("/subjects/{userId}") // 특정 사용자의 과목 리스트를 요청한다.
   public ResponseEntity<List<SubjectResponse>> getSubjectList(@PathVariable Long userId) {
@@ -71,7 +71,7 @@ public class StudyApiController {
   @GetMapping("/recodes/{userId}")
   public ResponseEntity<List<RecodeResponse>> getUserRecodes(@PathVariable Long userId,
       GetUserRecodesRequest request) {
-    User findUser = userUtil.findById(userId);
+    User findUser = userService.findById(userId);
     List<StudyRecord> records = studyService.getUserRecodes(findUser, request);
     List<RecodeResponse> recodeResponses = records.stream()
         .map(studyUtil::convertRecodeResponse)
