+++ backend/src/main/java/com/github/dadogk/user/mapper/UserResponseMapper.java
@@ -1,4 +1,4 @@
-package com.github.dadogk.user.util;
+package com.github.dadogk.user.mapper;
 
 import static com.github.dadogk.study.util.StudyUtil.calculateStudyTime;
 
@@ -6,8 +6,6 @@ import com.github.dadogk.study.entity.StudyRecord;
 import com.github.dadogk.study.entity.StudyRecordRepository;
 import com.github.dadogk.user.dto.UserResponse;
 import com.github.dadogk.user.entity.User;
-import com.github.dadogk.user.entity.UserRepository;
-import com.github.dadogk.user.exception.NotFoundUserException;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.util.List;
@@ -16,24 +14,9 @@ import org.springframework.stereotype.Component;
 
 @Component
 @AllArgsConstructor
-public class UserUtil {
+public class UserResponseMapper {
 
-  private final UserRepository userRepository;
   private final StudyRecordRepository studyRecordRepository;
-
-  /**
-   * @param userId User db id
-   * @return User
-   */
-  public User findById(Long userId) {
-    return userRepository.findById(userId)
-        .orElseThrow(() -> new NotFoundUserException("Unexpected user"));
-  }
-
-  public User findByEmail(String email) {
-    return userRepository.findByEmail(email)
-        .orElseThrow(() -> new NotFoundUserException("Unexpected user"));
-  }
 
   public UserResponse convertUserResponse(User findUser) {
     LocalDate nowDate = LocalDate.now();
@@ -42,7 +25,7 @@ public class UserUtil {
 
     List<StudyRecord> records = studyRecordRepository.findByUserAndStartAtBetween(findUser,
         startTime, endTime);
-    Long todayStudyTime = 0L;
+    long todayStudyTime = 0L;
     for (StudyRecord record : records) {
       todayStudyTime += calculateStudyTime(record);
     }
