+++ backend/src/test/java/com/github/dadogk/study/StudyApiControllerTest.java
@@ -14,7 +14,7 @@ import com.github.dadogk.study.entity.StudySubject;
 import com.github.dadogk.study.entity.StudySubjectRepository;
 import com.github.dadogk.user.entity.User;
 import com.github.dadogk.user.entity.UserRepository;
-import com.github.dadogk.user.util.TestUserUtil;
+import com.github.dadogk.user.mapper.TestUserUtil;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
